requirements.txt
dash==2.14.0
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.5
pandas
plotly==5.17.0
Pyrebase4==4.7.1
gunicorn==21.2.0
pvnvml==11.5.0
psutil==5.9.6
